# contracts/openapi/abfall-api.yaml
openapi: 3.1.0
info:
  title: Abfall APP API
  version: 0.2.0
  description: Contract-first API für den synthetischen, lokal testbaren Phase-1/2-Piloten.
  license:
    name: UNLICENSED
    identifier: LicenseRef-Proprietary
servers:
  - url: /
    description: Aktueller API-Ursprung; konkrete Basis-URLs kommen aus der Laufzeitkonfiguration.
security: []
...
paths:
...
  /v1/addresses/search:
    get:
      operationId: searchAddresses
      summary: Sucht belieferte Adressen eines Mandanten.
      tags: [Content]
      parameters:
        - $ref: "#/components/parameters/TenantId"
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 2
            maxLength: 120
      responses:
        "200":
          description: Passende synthetische Adressen.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Address"
        "400":
          $ref: "#/components/responses/BadRequest"
  /v1/addresses/{addressId}/collections:
    get:
      operationId: getCollections
      summary: Liefert kommende Abholtermine einer Adresse.
      tags: [Content]
      parameters:
        - $ref: "#/components/parameters/TenantId"
        - $ref: "#/components/parameters/AddressId"
      responses:
        "200":
          description: Chronologische Abholtermine.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CollectionEvent"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/waste-guide/search:
    get:
      operationId: searchWasteGuide
      summary: Durchsucht das Abfall-ABC fehlertolerant nach Namen und Synonymen.
      tags: [Content]
      parameters:
        - $ref: "#/components/parameters/TenantId"
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 2
            maxLength: 120
      responses:
        "200":
          description: Passende ABC-Einträge.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WasteGuideEntry"
        "400":
          $ref: "#/components/responses/BadRequest"
  /v1/sites:
    get:
      operationId: getSites
      summary: Listet Entsorgungsstandorte und optionale Abfallartfilter.
      tags: [Content]
      parameters:
        - $ref: "#/components/parameters/TenantId"
        - name: wasteType
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Filterbare Standortliste.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Site"
        "400":
          $ref: "#/components/responses/BadRequest"
  /v1/notices:
    get:
      operationId: getNotices
      summary: Liefert aktuell gültige Meldungen.
      tags: [Content]
      parameters:
        - $ref: "#/components/parameters/TenantId"
        - name: addressId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Aktive Meldungen nach Priorität.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notice"
        "400":
          $ref: "#/components/responses/BadRequest"
  /v1/admin/collections:
    get:
      operationId: getAdminCollections
      summary: Listet alle synthetischen Abholtermine zur Pflege.
      tags: [Admin]
      parameters: [{ $ref: "#/components/parameters/TenantId" }]
      responses:
        "200":
          description: Vollständige Terminliste.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/CollectionEvent" }
        "400": { $ref: "#/components/responses/BadRequest" }
    post:
      operationId: createCollection
      summary: Legt einen synthetischen Abholtermin an.
      tags: [Admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CollectionInput"
      responses:
        "201":
          description: Angelegter Termin.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionEvent"
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /v1/admin/waste-guide:
    get:
      operationId: getAdminWasteGuide
      summary: Listet alle synthetischen ABC-Einträge zur Pflege.
      tags: [Admin]
      parameters: [{ $ref: "#/components/parameters/TenantId" }]
      responses:
        "200":
          description: Vollständiges Abfall-ABC.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/WasteGuideEntry" }
        "400": { $ref: "#/components/responses/BadRequest" }
    post:
      operationId: createWasteGuideEntry
      summary: Legt einen synthetischen ABC-Eintrag an.
      tags: [Admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WasteGuideInput"
      responses:
        "201":
          description: Angelegter ABC-Eintrag.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WasteGuideEntry"
        "400":
          $ref: "#/components/responses/BadRequest"
  /v1/admin/sites:
    get:
      operationId: getAdminSites
      summary: Listet alle synthetischen Standorte zur Pflege.
      tags: [Admin]
      parameters: [{ $ref: "#/components/parameters/TenantId" }]
      responses:
        "200":
          description: Vollständige Standortliste.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Site" }
        "400": { $ref: "#/components/responses/BadRequest" }
    post:
      operationId: createSite
      summary: Legt einen synthetischen Entsorgungsstandort an.
      tags: [Admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SiteInput"
      responses:
        "201":
          description: Angelegter Standort.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Site"
        "400":
          $ref: "#/components/responses/BadRequest"
  /v1/admin/notices:
    get:
      operationId: getAdminNotices
      summary: Listet aktive, künftige und abgelaufene Hinweise zur Pflege.
      tags: [Admin]
      parameters: [{ $ref: "#/components/parameters/TenantId" }]
      responses:
        "200":
          description: Vollständige Hinweisliste.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Notice" }
        "400": { $ref: "#/components/responses/BadRequest" }
    post:
      operationId: createNotice
      summary: Legt eine synthetische Meldung an.
      tags: [Admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoticeInput"
      responses:
        "201":
          description: Angelegte Meldung.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notice"
        "400":
          $ref: "#/components/responses/BadRequest"
  /v1/admin/collections/{id}:
    put:
      operationId: updateCollection
      summary: Korrigiert einen synthetischen Abholtermin.
      tags: [Admin]
      parameters: [{ $ref: "#/components/parameters/ContentId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CollectionInput" }
      responses:
        "200":
          description: Aktualisierter Termin.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CollectionEvent" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      operationId: deleteCollection
      summary: Löscht einen synthetischen Abholtermin.
      tags: [Admin]
      parameters:
        - { $ref: "#/components/parameters/ContentId" }
        - { $ref: "#/components/parameters/TenantId" }
      responses:
        "204": { description: Termin gelöscht. }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/admin/waste-guide/{id}:
    put:
      operationId: updateWasteGuideEntry
      summary: Korrigiert einen synthetischen ABC-Eintrag.
      tags: [Admin]
      parameters: [{ $ref: "#/components/parameters/ContentId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WasteGuideInput" }
      responses:
        "200":
          description: Aktualisierter ABC-Eintrag.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WasteGuideEntry" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      operationId: deleteWasteGuideEntry
      summary: Löscht einen synthetischen ABC-Eintrag.
      tags: [Admin]
      parameters:
        - { $ref: "#/components/parameters/ContentId" }
        - { $ref: "#/components/parameters/TenantId" }
      responses:
        "204": { description: ABC-Eintrag gelöscht. }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/admin/sites/{id}:
    put:
      operationId: updateSite
      summary: Korrigiert einen synthetischen Standort.
      tags: [Admin]
      parameters: [{ $ref: "#/components/parameters/ContentId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SiteInput" }
      responses:
        "200":
          description: Aktualisierter Standort.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Site" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      operationId: deleteSite
      summary: Löscht einen synthetischen Standort.
      tags: [Admin]
      parameters:
        - { $ref: "#/components/parameters/ContentId" }
        - { $ref: "#/components/parameters/TenantId" }
      responses:
        "204": { description: Standort gelöscht. }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/admin/notices/{id}:
    put:
      operationId: updateNotice
      summary: Korrigiert einen synthetischen Hinweis.
      tags: [Admin]
      parameters: [{ $ref: "#/components/parameters/ContentId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/NoticeInput" }
      responses:
        "200":
          description: Aktualisierter Hinweis.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Notice" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      operationId: deleteNotice
      summary: Löscht einen synthetischen Hinweis.
      tags: [Admin]
      parameters:
        - { $ref: "#/components/parameters/ContentId" }
        - { $ref: "#/components/parameters/TenantId" }
      responses:
        "204": { description: Hinweis gelöscht. }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /v1/recycling-access/requests:
    post:
      operationId: createRecyclingAccessRequest
      summary: Beantragt einen synthetischen, zeitgebundenen Zugang zum Demo-Recyclinghof.
      description: >-
        Pilotendpunkt ohne reale Kennzeichenerkennung oder Schrankenansteuerung. Kennzeichen
        müssen mit DEMO- beginnen; alle Scanner- und Torereignisse sind Simulationen.
      tags: [Recycling access]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecyclingAccessRequestInput"
      responses:
        "201":
          description: Synthetische Zufahrtsberechtigung wurde ausgestellt.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecyclingAccessRequest"
        "200":
          description: Bereits idempotent angelegte Zufahrtsberechtigung; der Zugangscode wird nicht erneut ausgegeben.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecyclingAccessRequest"
        "400":
          $ref: "#/components/responses/BadRequest"
  /v1/recycling-access/requests/{reference}/simulation-events:
    post:
      operationId: simulateRecyclingAccessEvent
      summary: Führt genau ein synthetisches Scanner- oder Torereignis aus.
      description: >-
        Ausschließlich für den lokalen Showcase. Eine reale Hardwareintegration benötigt einen
        authentifizierten Gerätekanal, Safety-Sensorik und ein separates Threat Model.
      tags: [Recycling access]
      parameters:
        - name: reference
          in: path
          required: true
          schema:
            type: string
            pattern: "^DEMO-Z-[A-F0-9]{12}$"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecyclingAccessSimulationInput"
      responses:
        "200":
          description: Aktualisierter Zustand der synthetischen Zufahrt.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecyclingAccessRequest"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Das Ereignis passt nicht zum aktuellen Zustand.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
components:
  securitySchemes:
    monitoringToken:
      type: apiKey
      in: header
      name: X-Monitoring-Token
      description: Separates, mindestens 32 Zeichen langes Geheimnis für den Qualitätsagenten.
  parameters:
    ContentId:
      name: id
      in: path
      required: true
      schema:
        type: string
    TenantId:
      name: tenantId
      in: query
      required: true
      schema:
        type: string
        pattern: "^[a-z0-9][a-z0-9-]{0,62}$"
    AddressId:
      name: addressId
      in: path
      required: true
      schema:
        type: string
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 16
        maxLength: 120
  responses:
    BadRequest:
      description: Eingaben sind unvollständig oder fachlich ungültig.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    NotFound:
      description: Die angeforderte Ressource ist nicht vorhanden.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
  schemas:
...
    Address:
      type: object
      additionalProperties: false
      required: [id, tenantId, street, houseNumber, postalCode, city, displayLabel]
      properties:
        id:
          type: string
        tenantId:
          type: string
        street:
          type: string
        houseNumber:
          type: string
        postalCode:
          type: string
        city:
          type: string
        district:
          type: [string, "null"]
        displayLabel:
          type: string
    CollectionStatus:
      type: string
      enum: [planned, moved, cancelled, additional]
    CollectionEvent:
      type: object
      additionalProperties: false
      required:
        [
          id,
          addressId,
          wasteTypeId,
          wasteTypeLabel,
          plannedDate,
          effectiveDate,
          status,
          lastModified,
        ]
      properties:
        id:
          type: string
        addressId:
          type: string
        wasteTypeId:
          type: string
        wasteTypeLabel:
          type: string
        plannedDate:
          type: string
          format: date
        effectiveDate:
          type: string
          format: date
        status:
          $ref: "#/components/schemas/CollectionStatus"
        lastModified:
          type: string
          format: date-time
    CollectionInput:
      type: object
      additionalProperties: false
      required:
        [tenantId, addressId, wasteTypeId, wasteTypeLabel, plannedDate, effectiveDate, status]
      properties:
        tenantId:
          type: string
        addressId:
          type: string
        wasteTypeId:
          type: string
        wasteTypeLabel:
          type: string
          minLength: 2
        plannedDate:
          type: string
          format: date
        effectiveDate:
          type: string
          format: date
        status:
          $ref: "#/components/schemas/CollectionStatus"
    WasteGuideEntry:
      type: object
      additionalProperties: false
      required: [id, tenantId, name, category, disposalRoute, notes, synonyms, dataStatus]
      properties:
        id:
          type: string
        tenantId:
          type: string
        name:
          type: string
        category:
          type: string
        disposalRoute:
          type: string
        notes:
          type: string
        synonyms:
          type: array
          items:
            type: string
        dataStatus:
          type: string
          format: date-time
    WasteGuideInput:
      type: object
      additionalProperties: false
      required: [tenantId, name, category, disposalRoute, notes, synonyms]
      properties:
        tenantId:
          type: string
        name:
          type: string
          minLength: 2
        category:
          type: string
          minLength: 2
        disposalRoute:
          type: string
          minLength: 2
        notes:
          type: string
        synonyms:
          type: array
          items:
            type: string
    Site:
      type: object
      additionalProperties: false
      required:
        [
          id,
          tenantId,
          name,
          siteType,
          address,
          openingHours,
          acceptedWasteTypes,
          openNow,
          latitude,
          longitude,
          dataStatus,
        ]
      properties:
        id:
          type: string
        tenantId:
          type: string
        name:
          type: string
        siteType:
          type: string
        address:
          type: string
        openingHours:
          type: string
        acceptedWasteTypes:
          type: array
          items:
            type: string
        openNow:
          type: boolean
        latitude:
          type: number
          format: double
          minimum: -90
          maximum: 90
        longitude:
          type: number
          format: double
          minimum: -180
          maximum: 180
        dataStatus:
          type: string
          format: date-time
    SiteInput:
      type: object
      additionalProperties: false
      required:
        [
          tenantId,
          name,
          siteType,
          address,
          openingHours,
          acceptedWasteTypes,
          openNow,
          latitude,
          longitude,
        ]
      properties:
        tenantId:
          type: string
        name:
          type: string
          minLength: 2
        siteType:
          type: string
        address:
          type: string
        openingHours:
          type: string
        acceptedWasteTypes:
          type: array
          minItems: 1
          items:
            type: string
        openNow:
          type: boolean
        latitude:
          type: number
          format: double
          minimum: -90
          maximum: 90
        longitude:
          type: number
          format: double
          minimum: -180
          maximum: 180
    NoticePriority:
      type: string
      enum: [info, warning, critical]
    Notice:
      type: object
      additionalProperties: false
      required: [id, tenantId, noticeType, title, body, priority, validFrom, validUntil]
      properties:
        id:
          type: string
        tenantId:
          type: string
        addressId:
          type: [string, "null"]
        noticeType:
          type: string
        title:
          type: string
        body:
          type: string
        priority:
          $ref: "#/components/schemas/NoticePriority"
        validFrom:
          type: string
          format: date-time
        validUntil:
          type: string
          format: date-time
    NoticeInput:
      type: object
      additionalProperties: false
      required: [tenantId, noticeType, title, body, priority, validFrom, validUntil]
      properties:
        tenantId:
          type: string
        addressId:
          type: [string, "null"]
        noticeType:
          type: string
        title:
          type: string
          minLength: 3
        body:
          type: string
          minLength: 3
        priority:
          $ref: "#/components/schemas/NoticePriority"
        validFrom:
          type: string
          format: date-time
        validUntil:
          type: string
          format: date-time
...
    RecyclingAccessIdentificationMethod:
      type: string
      enum: [code, license-plate]
    RecyclingAccessStatus:
      type: string
      enum: [authorized, entry-granted, on-site, exit-granted, completed]
    RecyclingGateState:
      type: string
      enum: [closed, open-entry, open-exit]
    RecyclingAccessSimulationEventType:
      type: string
      enum: [arrival-scan, entry-confirmed, exit-scan, exit-confirmed]
    RecyclingAccessRequestInput:
      type: object
      additionalProperties: false
      required:
        [
          tenantId,
          siteId,
          plannedArrivalAt,
          wasteType,
          itemDescription,
          identificationMethod,
          syntheticDataConfirmed,
        ]
      properties:
        tenantId:
          type: string
        siteId:
          type: string
        plannedArrivalAt:
          type: string
          format: date-time
        wasteType:
          type: string
          enum: [electronics]
        itemDescription:
          type: string
          minLength: 3
          maxLength: 160
        identificationMethod:
          $ref: "#/components/schemas/RecyclingAccessIdentificationMethod"
        syntheticLicensePlate:
          type: [string, "null"]
          pattern: "^DEMO-[A-Z0-9-]{2,20}$"
        syntheticDataConfirmed:
          type: boolean
    RecyclingAccessSimulationInput:
      type: object
      additionalProperties: false
      required: [accessToken, eventType, credential]
      properties:
        accessToken:
          type: string
          minLength: 32
        eventType:
          $ref: "#/components/schemas/RecyclingAccessSimulationEventType"
        credential:
          type: string
          minLength: 4
          maxLength: 32
    RecyclingAccessEvent:
      type: object
      additionalProperties: false
      required: [eventType, label, occurredAt]
      properties:
        eventType:
          type: string
        label:
          type: string
        occurredAt:
          type: string
          format: date-time
    RecyclingAccessRequest:
      type: object
      additionalProperties: false
      required:
        [
          reference,
          accessToken,
          tenantId,
          siteId,
          siteName,
          plannedArrivalAt,
          accessWindowStart,
          accessWindowEnd,
          wasteType,
          itemDescription,
          identificationMethod,
          credentialHint,
          status,
          gateState,
          nextSimulationEvent,
          events,
          createdAt,
          updatedAt,
        ]
      properties:
        reference:
          type: string
          pattern: "^DEMO-Z-[A-F0-9]{12}$"
        accessToken:
          type: string
        tenantId:
          type: string
        siteId:
          type: string
        siteName:
          type: string
        plannedArrivalAt:
          type: string
          format: date-time
        accessWindowStart:
          type: string
          format: date-time
        accessWindowEnd:
          type: string
          format: date-time
        wasteType:
          type: string
        itemDescription:
          type: string
        identificationMethod:
          $ref: "#/components/schemas/RecyclingAccessIdentificationMethod"
        credential:
          type: [string, "null"]
          description: Wird nur bei der erstmaligen Ausstellung zurückgegeben.
        credentialHint:
          type: string
        status:
          $ref: "#/components/schemas/RecyclingAccessStatus"
        gateState:
          $ref: "#/components/schemas/RecyclingGateState"
        nextSimulationEvent:
          type: [string, "null"]
        events:
          type: array
          items:
            $ref: "#/components/schemas/RecyclingAccessEvent"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Problem:
      type: object
      additionalProperties: true
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          format: uri-reference